Create app.js
This JavaScript file fetches recipes from the Edamam API based on a user input search query and displays the recipe information on the web page. When the user clicks the "Search" button, the input value is used to make a fetch request to the Edamam API, and the response data is used to generate recipe cards that are displayed on the page. Each recipe card contains an image, title, description, and a link to the recipe page. If an error occurs while fetching the recipes, an error message is displayed to the user.